+++ packages/integration-sdk-testing/src/__tests__/jest.test.ts
@@ -597,6 +597,68 @@ Find out more about JupiterOne schemas: https://github.com/JupiterOne/data-model
 Find out more about JupiterOne schemas: https://github.com/JupiterOne/data-model/tree/main/src/schemas
 `);
   });
+
+  test('should require matching `_type` if `_type` is provided', () => {
+    const relationship = generateCollectedDirectRelationship({
+      string: 'abc',
+      number: 123,
+      boolean: true,
+      null: null,
+      _type: 'relationship-type',
+    });
+
+    const result = toMatchDirectRelationshipSchema(relationship, {
+      _type: 'different-relationship-type',
+    });
+
+    expect(result).toEqual({
+      message: expect.any(Function),
+      pass: false,
+    });
+
+    expect(result.message()).toMatch(`errors=[
+  {
+    "instancePath": "/_type",
+    "schemaPath": "#/properties/_type/const",
+    "keyword": "const",
+    "params": {
+      "allowedValue": "different-relationship-type"
+    },
+    "message": "must be equal to constant"
+  }
+]`);
+  });
+
+  test('should require matching `_class` if `_class` is provided', () => {
+    const relationship = generateCollectedDirectRelationship({
+      string: 'abc',
+      number: 123,
+      boolean: true,
+      null: null,
+      _class: RelationshipClass.HAS,
+    });
+
+    const result = toMatchDirectRelationshipSchema(relationship, {
+      _class: RelationshipClass.IS,
+    });
+
+    expect(result).toEqual({
+      message: expect.any(Function),
+      pass: false,
+    });
+
+    expect(result.message()).toMatch(`errors=[
+  {
+    "instancePath": "/_class",
+    "schemaPath": "#/properties/_class/const",
+    "keyword": "const",
+    "params": {
+      "allowedValue": "IS"
+    },
+    "message": "must be equal to constant"
+  }
+]`);
+  });
 });
 
 describe('#toTargetEntities', () => {