+++ CHANGELOG.md
@@ -11,6 +11,30 @@ and this project adheres to
 
 ### Added
 
+- Added `executeStepWithDependencies` utility to
+  `@jupiterone/integration-sdk-testing` package. This allows developers to test
+  specific integration steps in isolation, while assuring that all of its
+  dependencies have indeed executed. Usage:
+
+  ```ts
+  const {
+    collectedEntities,
+    collectedRelationships,
+    collectedData,
+  } = await executeStepWithDependencies({
+    stepId: Steps.FETCH_USERS.id,
+    invocationConfig,
+    instanceConfig,
+  });
+
+  expect(collectedEntities.length).toBeGreaterThan(0);
+  expect(collectedEnities).toMatchGraphObjectSchema({
+    _class: Entities.USER._class,
+    schema: Entities.USER.schema,
+  });
+  // ... additional expectations
+  ```
+
 - Added `MockJobState.collectedData` to capture data that has been collected in
   the job state. Usage: